backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy>=2.0.35
python-dotenv==1.0.1
python-multipart>=0.0.9

# NLP & Text Processing
textstat==0.7.3
spacy>=3.8.0
python-docx==1.1.2
pypdf>=4.0.0
nltk==3.9.1

# AI & LLM
openai>=1.0.0
google-genai>=1.0.0

# Audio / TTS
gtts==2.5.1